+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   * Repository label updated to display URL and package name.
   * Exclude flags updated to hide inaccessible fields in access management.
   * Hook changed to update non-defined repository package on every save.
+  * Package includes changed to reflect the package name instead of the repository UID.
 
 ## [1.0.0] - 2021-11-06
 Initial version committed.